Fix critical bug that if you don't use LoadWorld or Instance, you may delete the world the quest is on.

+++ src/main/java/com/theminequest/MineQuest/Quest/QuestWorldManip.java
@@ -51,8 +51,8 @@ public class QuestWorldManip {
 	
 	public static void removeWorld(World w) throws IOException{
 		String removename = w.getName();
-		Bukkit.unloadWorld(w, false);
-		FileUtils.deleteDirectory(new File(removename));
+		if (Bukkit.unloadWorld(w, false))
+			FileUtils.deleteDirectory(new File(removename));
 	}
 	
 }
